requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.6.2=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-generator=1.10=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blinker=1.5=pypi_0
bottleneck=1.3.5=py37h67323c0_0
brotli=1.0.9=pypi_0
brotlipy=0.7.0=py37h9ed2024_1003
ca-certificates=2022.10.11=hecd8cb5_0
certifi=2022.9.24=py37hecd8cb5_0
cffi=1.15.1=py37hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=38.0.1=py37hf6deb26_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.0.0=pypi_0
fastjsonschema=2.16.2=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.4=py37hecd8cb5_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.10.0=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.16.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.4=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=4.11.2=pypi_0
jupyter-server=1.21.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.3=pypi_0
kaitaistruct=0.10=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
markupsafe=2.1.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py37h9ed2024_0
mkl_fft=1.3.1=py37h4ab4a9b_0
mkl_random=1.2.2=py37hb2f4e1b_0
nbclassic=0.4.7=pypi_0
nbclient=0.7.0=pypi_0
nbconvert=7.2.3=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.6=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.0=pypi_0
numexpr=2.8.3=py37h2e5f0a9_0
numpy=1.21.5=py37h2e5f0a9_3
numpy-base=1.21.5=py37h3b1a694_3
openssl=1.1.1q=hca72f7f_0
outcome=1.2.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h743cdd8_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=22.2.2=py37hecd8cb5_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.15.0=pypi_0
prompt-toolkit=3.0.31=pypi_0
psutil=5.9.3=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37hecd8cb5_0
pyrsistent=0.19.1=pypi_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.13=hdfd78df_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=pypi_0
pytz=2022.1=py37hecd8cb5_0
pyzmq=24.0.1=pypi_0
qtconsole=5.3.2=pypi_0
qtpy=2.2.1=pypi_0
readline=8.2=hca72f7f_0
requests=2.28.1=py37hecd8cb5_0
schedule=1.1.0=pypi_0
selenium=4.5.0=pypi_0
selenium-wire=5.1.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=63.4.1=py37hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.3=h707629a_0
terminado=0.17.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h5d9f67b_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
trio=0.22.0=pypi_0
trio-websocket=0.9.2=pypi_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.12=py37hecd8cb5_0
wcwidth=0.2.5=pypi_0
webdriver-manager=3.8.4=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pypi_0
wsproto=1.2.0=pypi_0
xz=5.2.6=hca72f7f_0
zipp=3.10.0=pypi_0
zlib=1.2.13=h4dc903c_0
zstandard=0.19.0=pypi_0
